+++ src/lib.rs
@@ -41,7 +41,7 @@ pub fn ffprobe_config(
     let mut cmd = std::process::Command::new("ffprobe");
 
     // Default args.
-    cmd.args(&[
+    cmd.args([
         "-v",
         "quiet",
         "-show_format",
@@ -110,6 +110,12 @@ impl ConfigBuilder {
     /// Run ffprobe with the config produced by this builder.
     pub fn run(self, path: impl AsRef<std::path::Path>) -> Result<FfProbe, FfProbeError> {
         ffprobe_config(self.config, path)
+    }
+}
+
+impl Default for ConfigBuilder {
+    fn default() -> Self {
+        Self::new()
     }
 }
 
@@ -205,12 +211,16 @@ pub struct Stream {
 
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(feature = "__internal_deny_unknown_fields", serde(deny_unknown_fields))]
+// Allowed to prevent having to break compatibility of float fields are added.
+#[allow(clippy::derive_partial_eq_without_eq)]
 pub struct SideData {
     pub side_data_type: String,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(feature = "__internal_deny_unknown_fields", serde(deny_unknown_fields))]
+// Allowed to prevent having to break compatibility of float fields are added.
+#[allow(clippy::derive_partial_eq_without_eq)]
 pub struct Disposition {
     pub default: i64,
     pub dub: i64,
@@ -228,6 +238,8 @@ pub struct Disposition {
 
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(feature = "__internal_deny_unknown_fields", serde(deny_unknown_fields))]
+// Allowed to prevent having to break compatibility of float fields are added.
+#[allow(clippy::derive_partial_eq_without_eq)]
 pub struct StreamTags {
     pub language: Option<String>,
     pub creation_time: Option<String>,
@@ -277,6 +289,7 @@ impl Format {
 
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(feature = "__internal_deny_unknown_fields", serde(deny_unknown_fields))]
+#[allow(clippy::derive_partial_eq_without_eq)]
 pub struct FormatTags {
     #[serde(rename = "WMFSDKNeeded")]
     pub wmfsdkneeded: Option<String>,